Persist and display sub task cost

+++ presentation/src/main/java/szysz3/planty/screen/tasklist/model/SubTask.kt
@@ -4,5 +4,5 @@ data class SubTask(
     val id: Long = 0,
     val description: String = "",
     val isCompleted: Boolean = false,
-    val cost: Int = 0
+    val cost: Float = 0f
 )
